TQP `int` -> `long` for vsId

+++ TaskQueueProcessor/src/NIST.CVP.TaskQueueProcessor/Providers/IJsonProvider.cs
@@ -2,7 +2,7 @@ namespace NIST.CVP.TaskQueueProcessor.Providers
 {
     public interface IJsonProvider
     {
-        string GetJson(int vsId, string jsonFileType);
-        void PutJson(int vsId, string jsonFileType, string json);
+        string GetJson(long vsId, string jsonFileType);
+        void PutJson(long vsId, string jsonFileType, string json);
     }
 }
